pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cloudzipfile"
authors = [{name = "Jesse Wursten", email = "jessaius+cloudzipfile@hotmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">=3.0"
dependencies = [
    "azure-storage-blob"
]

[project.urls]
Source = "https://github.com/Danferno/cloudzipfile"

[project.optional-dependencies]
dev = [
    "flit"
]